feat: 添加sql拦截器，统一对createdUser、createdTime、updatedUser、updatedTime字段进行处理

+++ renting-integration-dal/src/main/java/cn/waynechu/renting/dal/mapper/HouseMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface HouseMapper {
-    long countByExample(HouseExample example);
+    int countByExample(HouseExample example);
 
     int deleteByExample(HouseExample example);
